requirements.txt
# agentic
autogen-ext[langchain]
autogen-agentchat
autogen-ext[openai,azure]


# llm
langchain-openai
python-dotenv
langchain-chroma
langchain-community
certifi
httpx


# read_file.py
python-docx
PyPDF2
pdfminer.six


# ml_unstructured.py
numpy
scikit-learn
xgboost
lightgbm
catboost
umap-learn
